test: add environment variable `PROPTEST_CASES`
Move misc tests to wasm_generator
Remove leftover file `ast_visitor`
