+++ clar2wasm/src/wasm_generator.rs
@@ -1551,3 +1551,35 @@ impl WasmGenerator {
         }
     }
 }
+
+#[cfg(test)]
+mod misc_tests {
+    // Tests that don't relate to specific words
+    use crate::tools::{crosscheck, evaluate};
+
+    #[test]
+    fn is_in_mainnet() {
+        crosscheck(
+            "
+(define-public (mainnet)
+  (ok is-in-mainnet))
+
+(mainnet)
+",
+            evaluate("(ok false)"),
+        );
+    }
+
+    #[test]
+    fn is_in_regtest() {
+        crosscheck(
+            "
+(define-public (regtest)
+  (ok is-in-regtest))
+
+(regtest)
+",
+            evaluate("(ok false)"),
+        );
+    }
+}
