+++ clar2wasm/tests/wasm-generation/conditionals.rs
@@ -5,7 +5,7 @@ use proptest::strategy::{Just, Strategy};
 use crate::{prop_signature, type_string, PropValue};
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
     #[test]
     fn if_true_returns_first_value(
         (v1, v2) in prop_signature()
@@ -22,7 +22,7 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
     #[test]
     fn if_false_returns_second_value(
         (v1, v2) in prop_signature()
@@ -39,7 +39,7 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
     #[test]
     fn match_optional_some(
         (initial, some_val, none_val) in (prop_signature(), prop_signature())
@@ -54,7 +54,6 @@ proptest! {
         )
     }
 
-    #[ignore]
     #[test]
     fn match_optional_none(
         (original_ty, some_val, none_val) in (prop_signature(), prop_signature())
@@ -72,7 +71,6 @@ proptest! {
         )
     }
 
-    #[ignore]
     #[test]
     fn match_response_ok(
         (original_ok_ty, original_ok_val, original_err_ty, ok_val, err_val) in (prop_signature(), prop_signature(), prop_signature())
@@ -89,7 +87,7 @@ proptest! {
             Ok(Some(ok_val.into()))
         )
     }
-    #[ignore]
+
     #[test]
     fn match_response_err(
         (original_ok_ty, original_err_ty, original_err_val, ok_val, err_val) in (prop_signature(), prop_signature(), prop_signature())
