+++ clar2wasm/tests/wasm-generation/equal.rs
@@ -4,7 +4,8 @@ use proptest::proptest;
 use crate::PropValue;
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_eq_one_argument_always_true(val in PropValue::any()) {
         crosscheck(
@@ -15,7 +16,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_eq_value_with_itself_always_true(val in PropValue::any()) {
         crosscheck(
@@ -26,7 +28,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_eq_value_with_itself_always_true_3(val in PropValue::any()) {
         crosscheck(
