+++ clar2wasm/tests/wasm-generation/optional.rs
@@ -4,7 +4,8 @@ use proptest::proptest;
 use crate::PropValue;
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_some_always_true(val in PropValue::any()) {
         crosscheck(
@@ -15,7 +16,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_none_always_false(val in PropValue::any()) {
         crosscheck(
