+++ clar2wasm/tests/wasm-generation/response.rs
@@ -4,7 +4,8 @@ use proptest::proptest;
 use crate::PropValue;
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_ok_always_true(val in PropValue::any()) {
         crosscheck(
@@ -15,7 +16,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_ok_always_false(val in PropValue::any()) {
         crosscheck(
@@ -26,7 +28,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_err_always_true(val in PropValue::any()) {
         crosscheck(
@@ -37,7 +40,8 @@ proptest! {
 }
 
 proptest! {
-    #[ignore]
+    #![proptest_config(super::runtime_config())]
+
     #[test]
     fn is_err_always_false(val in PropValue::any()) {
         crosscheck(
